@@ -12,39 +12,50 @@
     "ionic:serve": "ionic-app-scripts serve"
   },
   "dependencies": {
-    "@angular/common": "5.0.3",
-    "@angular/compiler": "5.0.3",
-    "@angular/compiler-cli": "5.0.3",
-    "@angular/core": "5.0.3",
-    "@angular/forms": "5.0.3",
-    "@angular/http": "5.0.3",
-    "@angular/platform-browser": "5.0.3",
-    "@angular/platform-browser-dynamic": "5.0.3",
+    "@angular/common": "6.1.4",
+    "@angular/compiler": "6.1.4",
+    "@angular/compiler-cli": "6.1.4",
+    "@angular/core": "6.1.4",
+    "@angular/forms": "6.1.4",
+    "@angular/http": "6.1.4",
+    "@angular/platform-browser": "6.1.4",
+    "@angular/platform-browser-dynamic": "6.1.4",
+    "@ionic-native/barcode-scanner": "^4.12.2",
     "@ionic-native/bluetooth-serial": "^4.5.2",
-    "@ionic-native/core": "4.4.0",
+    "@ionic-native/core": "4.12.0",
+    "@ionic-native/dialogs": "^4.6.0",
+    "@ionic-native/qr-scanner": "^4.12.2",
     "@ionic-native/screen-orientation": "^4.5.2",
-    "@ionic-native/splash-screen": "4.4.0",
+    "@ionic-native/splash-screen": "4.12.0",
     "@ionic-native/status-bar": "^4.4.0",
-    "@ionic/storage": "2.1.3",
-    "cordova-android": "6.3.0",
+    "@ionic/storage": "^2.1.3",
+    "android-versions": "^1.3.0",
+    "angularx-qrcode": "^1.2.2",
+    "cordova-android": "7.0.0",
     "cordova-plugin-bluetooth-serial": "^0.4.7",
-    "cordova-plugin-device": "^1.1.4",
-    "cordova-plugin-ionic-webview": "^1.1.16",
+    "cordova-plugin-device": "^2.0.2",
+    "cordova-plugin-dialogs": "^2.0.1",
+    "cordova-plugin-ionic-webview": "^2.1.3",
     "cordova-plugin-screen-orientation": "^3.0.1",
-    "cordova-plugin-splashscreen": "^4.0.3",
-    "cordova-plugin-statusbar": "^2.4.1",
-    "cordova-plugin-whitelist": "^1.3.1",
+    "cordova-plugin-splashscreen": "^5.0.2",
+    "cordova-plugin-statusbar": "^2.4.2",
+    "cordova-plugin-whitelist": "^1.3.3",
+    "cordova-sqlite-storage": "^2.4.0",
     "es6-promise-plugin": "^4.2.2",
     "ionic-angular": "3.9.2",
     "ionic-plugin-keyboard": "^2.2.1",
-    "ionicons": "3.0.0",
-    "rxjs": "5.5.2",
+    "ionicons": "4.4.3",
+    "lzutf8": "^0.5.5",
+    "phonegap-plugin-barcodescanner": "^8.0.0",
+    "rxjs": "6.2.2",
+    "rxjs-compat": "^6.2.2",
     "sw-toolbox": "3.6.0",
-    "zone.js": "0.8.18"
+    "webcom": "^2.5.0",
+    "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@ionic/app-scripts": "3.1.6",
-    "typescript": "2.4.2"
+    "@ionic/app-scripts": "3.2.0",
+    "typescript": "3.0.1"
   },
   "description": "An Ionic project",
   "cordova": {
@@ -56,7 +67,12 @@
       "cordova-plugin-splashscreen": {},
       "cordova-plugin-ionic-webview": {},
       "cordova-plugin-screen-orientation": {},
-      "cordova-plugin-statusbar": {}
+      "cordova-plugin-statusbar": {},
+      "cordova-plugin-dialogs": {},
+      "cordova-sqlite-storage": {},
+      "phonegap-plugin-barcodescanner": {
+        "ANDROID_SUPPORT_V4_VERSION": "27.+"
+      }
     },
     "platforms": [
       "android"
