@@ -3,7 +3,9 @@ import { Platform } from 'ionic-angular';
 import { Injectable } from '@angular/core';
 import { BluetoothSerial } from '@ionic-native/bluetooth-serial';
 import { Storage } from '@ionic/storage';
-import { Observable } from 'rxjs/Observable';
+import { ShareProvider } from '../share/share';
+
+//import { Observable } from 'rxjs/Observable';
 /*
   Generated class for the BluetoothProvider provider.
 
@@ -15,11 +17,8 @@ import { Observable } from 'rxjs/Observable';
 @Injectable()
 export class BluetoothProvider {
 
-  constructor(private bluetoothSerial: BluetoothSerial, private storage: Storage, private platform: Platform) {
+  constructor(private bluetoothSerial: BluetoothSerial, private shareProvider: ShareProvider, private storage: Storage, private platform: Platform) {
     console.log('Hello BluetoothProvider Provider');
-
-
-
   }
   public btledevices;
 
@@ -29,7 +28,6 @@ export class BluetoothProvider {
 
   setPreferedDevice(device) {
     return this.storage.set("device", device);
-
   }
 
   getPreferedDevice() {
@@ -37,11 +35,11 @@ export class BluetoothProvider {
   }
 
   disconnect() {
+    this.shareProvider.disconnect();
     return this.bluetoothSerial.disconnect();
   }
 
   serialWritePreferedDevice(command) {
-
     console.log("BT serial send :" + command)
 
     return new Promise((resolve, reject) => {
@@ -56,7 +54,6 @@ export class BluetoothProvider {
         }
         else {
           this.getPreferedDevice().then((device) => {
-
             this.connectDevice(device).subscribe((data) => {
               this.bluetoothSerial.write(command).then(() => {
                 resolve();
@@ -69,17 +66,14 @@ export class BluetoothProvider {
 
           })
         }
-
-
       })
-
-
     })
-
   }
 
   connectDevice(device) {
-    return this.bluetoothSerial.connect(device.id);
+    const deviceId=device.id;
+    //this.shareProvider.shareRobot(deviceId);
+    return this.bluetoothSerial.connect(deviceId);
   }
 
 }
